pyproject.toml
[project]
name = "upstage-stock-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.4.1",
    "fastapi>=0.128.0",
    "finance-datareader>=0.9.101",
    "finance-tools>=0.1",
    "langchain>=1.2.3",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.6",
    "langchain-openai>=1.1.7",
    "langchain-tavily>=0.2.16",
    "langchain-upstage>=0.7.5",
    "lxml>=6.0.2",
    "opendartreader>=0.2.3",
    "pandas>=2.3.3",
    "pandas-stubs~=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "pypdf>=4.3.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "tavily-python>=0.7.17",
    "tiktoken>=0.12.0",
    "tools>=1.0.17",
    "uvicorn>=0.40.0",
    "yfinance>=1.0",
]
